+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - `fxid()` specific fx:id name as optional parameter 
 
+### Changed
+
+- Call `onUndock` when UIComponent.modalStage closes
+
 ## [1.5.0] - 2016-06-10
 
 ### Added